+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - User field to configure log notifications.
 - Add static scopes for rothamsted_data_admin and rothamsted_farm_manager.
+- Add rothamsted API consumer.
 
 ### Changed